rename to use pkg/v1 and pkg/v2
