+++ pkg/regression/v1_login.go
@@ -3,7 +3,7 @@
 package regression
 
 import (
-	storageapi "github.com/Seagate/seagate-exos-x-api-go/pkg/exosx"
+	storageapi "github.com/Seagate/seagate-exos-x-api-go/pkg/v1"
 	"k8s.io/klog/v2"
 
 	. "github.com/onsi/ginkgo/v2"
