+++ pkg/regression/v1_system.go
@@ -3,7 +3,7 @@
 package regression
 
 import (
-	storageapi "github.com/Seagate/seagate-exos-x-api-go/pkg/exosx"
+	storageapi "github.com/Seagate/seagate-exos-x-api-go/pkg/v1"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"k8s.io/klog/v2"
